Return richer fragment information
The return value from expanding a file contains the fragments used
to build the response. This is now richer and contains the filename
searched for as a value not a key (which makes tools built on this
response easier to code), and the content of the file before it is
itself expanded.