Provide the dimension possibilities
It will be useful when providing tools that can create more or
less specific fragments to have access to the list of possible
dimensions and their values for a given target.

This adds a new function `get_dimension_possibilities` that does just
that, and refactors `get_fragment_paths` to use it rather than to have
similar logic in two places (which could lead to divergence).