Add interface for manipulating fragments
Although these are simple file operations, it is helpful to have
it in the Sectile library so that the client software doesn't need
to remember to combine the fragment directory with the fragment
filename, which is what it is dealing with.

Plus creating a fragment is just a little more involved than making
a file. The new fragment should be a copy of the less specific
fragment, if one exists.